+++ src/routes/blog/articles/[slug]/+page.svelte
@@ -2,7 +2,7 @@
 	import HeadMetadata from '$lib/components/HeadMetadata.svelte';
 	import Space from '$lib/components/Space.svelte';
 	import ShareButton from '$lib/components/ShareButton.svelte';
-	import FeedButton from '$lib/components/FeedButton.svelte';
+	import FeedButton from '$lib/components/blog/FeedButton.svelte';
 	import Article from './Article.svelte';
 	import TagList from '$lib/btpc/components/tag/TagList.svelte';
 	import ScrollToTop from './ScrollToTop.svelte';
